fix JDBC syntax & remove redundant coder
